+++ src/Entity/Transaction/KeyTransaction.php
@@ -21,7 +21,7 @@
 
 namespace Adshares\Ads\Entity\Transaction;
 
-use DateTime;
+use DateTimeInterface;
 
 /**
  * Transaction type=<'account_created', 'change_account_key', 'change_node_key'>.
@@ -79,7 +79,7 @@ class KeyTransaction extends AbstractTransaction
     protected $targetUser;
 
     /**
-     * @var DateTime
+     * @var DateTimeInterface
      */
     protected $time;
 
@@ -169,9 +169,9 @@ class KeyTransaction extends AbstractTransaction
     }
 
     /**
-     * @return DateTime
+     * @return DateTimeInterface
      */
-    public function getTime(): DateTime
+    public function getTime(): DateTimeInterface
     {
         return $this->time;
     }