+++ tests/Unit/Util/AdsNormalizerTest.php
@@ -29,9 +29,16 @@ class AdsNormalizerTest extends TestCase
 {
     public function testNormalizeAddressValid(): void
     {
+        $address = '0001000000001234';
+        $address = AdsNormalizer::normalizeAddress($address);
+        $this->assertEquals('0001-00000000-1234', $address);
+    }
+
+    public function testNormalizeAddressWithoutChecksum(): void
+    {
         $address = '000100000000XXXX';
         $address = AdsNormalizer::normalizeAddress($address);
-        $this->assertEquals('0001-00000000-XXXX', $address);
+        $this->assertEquals('0001-00000000-9B6F', $address);
     }
 
     public function testNormalizeAddressInvalid(): void